requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
fastapi==0.101.0
filelock==3.12.2
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
opencv-python==4.8.0.74
Pillow==10.0.0
pydantic==2.1.1
pydantic_core==2.4.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
torch==2.0.1
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==2.0.4
